requirements.txt
flet==0.22.0
pyrebase4==4.8.0